backend/requirements.txt
# Core Dependencies
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pydantic>=2.6.4

# Dev Tools
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
